Breaking: Hide nom error type from public API
In #77, we made a breaking change—but I messed up the commit messages so
`semantic-version` didn't catch that, my apologies!—to upgrade the
version of `nom` we're using. That change needed to be breaking because
we expose a `nom::Err` type in `Program::from_str` in the public API.

This PR makes a small change to remove that type from the public API so
that future updates involving `nom` are no longer breaking changes.
Instead, the error type of `Program::from_str` is now `String`, which
matches `Expression::from_str`.