fix: fix performance regression (#113)
* bench: add calibration benchmark

* refactor: apply referenced error wherever possible

The performance culprit was `crate::parser::error::error::Error`, which
was being created more often than expected and creating strings every
time. The fix is to have an error type used while parsing that contains
a reference to the input, and only change to the more expensive error
type when done parsing.

A bunch of parsing functions were marked `pub` but were implicitly
`pub(crate)` from being in private modules. These were explicitly marked
`pub(crate)` and modified to use the new error type.

* chore: remove outdated comment

* refactor: code cleanup

* chore: disable lint on parameterized tests

* chore!: update to Rust 2021

This is needed to fix an issue clippy found where a bunch of `panic!`
invocations use the names of variables in scope without assigning them
in the macro, e.g. `panic!("{foo}")` instead of `panic!("{foo}", foo=foo)`.

This is supported in Rust 2021, but not 2018.

* style: run cargo fmt

* chore: remove extra TODO

Co-authored-by: Mark Skilbeck <mark.skilbeck@rigetti.com>

* chore: address code review comments

* chore: resolve cargo-deny error, remove unused ignored advisory

* style: run cargo fmt

* fix: calculate token line/column *after* parsing success

* chore: fix typo in comment

Co-authored-by: Kalan <22137047+kalzoo@users.noreply.github.com>

Co-authored-by: Mark Skilbeck <mark.skilbeck@rigetti.com>
Co-authored-by: Kalan <22137047+kalzoo@users.noreply.github.com>