feat!: add named measurements (#479)
Adds an optional `!name` to `MEASURE` and `DEFCAL MEASURE
and updates the corresponding calibration logic.
As written, the change uses a careful Python signature
so that the constructor changes
are not actually a breaking change for `quil` users.
This additionally required implementing `__getnewargs_ex__`.
Finally, these changes exposed a small bug in the linter script,
wherein it missed `pyo3_stub_gen` usages when written with a full path.

---------

Co-authored-by: Alexander Saites <215926835+asaites@users.noreply.github.com>
Closes: #478

BREAKING CHANGES: `quil-rs` users will need to modify `Measurement` and `MeasureCalibrationIdentifier` struct and constructor usage. `quil` users are unaffected.