style: use same predicate semantics across API
Currently `filter_instructions` uses the name "predicate",
an ordinary Python `Callable`, and bubbles errors as `PanicException`s.
This edits the new defgate functions to use the same naming and semantics.

It also introduces a `call_predicate` function
to simplify the common logic associated with calling those predicates,
and refactors the relevant to code to make use of it.