Asaites/447 defgate as sequence (#487)
* style: use same predicate semantics across API

Currently `filter_instructions` uses the name "predicate",
an ordinary Python `Callable`, and bubbles errors as `PanicException`s.
This edits the new defgate functions to use the same naming and semantics.

It also introduces a `call_predicate` function
to simplify the common logic associated with calling those predicates,
and refactors the relevant to code to make use of it.

* style: refactor defgate seq expander with stack

* style: apply clippy suggestions

* chore: regenerate stubs

* style: remove line

* fix: self import for python 3.10

---------

Co-authored-by: Alexander Saites <215926835+asaites@users.noreply.github.com>